// src/utils/widget-enums.ts
export enum WIDGET_COLOR {
  BLUE = 'blue',
  GREEN = 'green',
  BEIGE = 'beige',
  WHITE = 'white',
  BLACK = 'black',
}

export enum IMPACT_TYPE {
  CARBON = 'carbon',
  PLASTIC = 'plastic', // according to endpoint
  PLASTIC_BOTTLES = 'plastic bottles',
  TREES = 'trees',
}

export enum ACTION_TYPE {
  COLLECTS = 'collects',
  PLANTS = 'plants',
  OFFSETS = 'offsets',
}